+++ DTO/SaleDetail.cs
@@ -5,9 +5,9 @@
         public string Barcode { get; set; }
         public string Name { get; set; }
         public string Unit { get; set; }
-        public float UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
         public int Number { get; set; }
-        public float Amount { get => UnitPrice * Number; }
+        public double Amount { get => UnitPrice * Number; }
         public SaleDetail() { }
         public SaleDetail(string barcode, string name, string unit, float unitPrice, int number)
         {
